requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
autoenv==1.0.0
click==6.7
coverage==4.5.1
coveralls==1.3.0
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
gunicorn==19.9.0
funcsigs==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.6.0
PyJWT==1.6.4
pylint==1.9.3
pytest==3.7.4
pytest-cov==2.5.1
pytz==2018.5
scandir==1.9.0
six==1.11.0
Werkzeug==0.14.1
